refactor: Order 도메인 내 패키지 분류
