--- src/main/java/kr/hhplus/be/server/order/OrderRepository.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.order;
-
-public interface OrderRepository {
-}
