--- src/main/java/kr/hhplus/be/server/order/OrderResponse.java
@@ -1,4 +1,0 @@
-package kr.hhplus.be.server.order;
-
-public class OrderResponse {
-}
