+++ src/main/java/kr/hhplus/be/server/order/application/OrderFacade.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.order.application;
+
+public class OrderFacade {
+}
