+++ src/main/java/kr/hhplus/be/server/order/application/OrderService.java
@@ -1,0 +1,16 @@
+package kr.hhplus.be.server.order.application;
+
+import kr.hhplus.be.server.common.exception.EntityNotFoundException;
+import kr.hhplus.be.server.order.domain.Order;
+import kr.hhplus.be.server.order.infrastructure.OrderRepository;
+import lombok.RequiredArgsConstructor;
+import org.springframework.stereotype.Service;
+import org.springframework.transaction.annotation.Transactional;
+
+@Service
+@Transactional(readOnly = true)
+@RequiredArgsConstructor
+public class OrderService {
+    private final OrderRepository orderRepository;
+
+}
