+++ src/main/java/kr/hhplus/be/server/order/domain/OrderItem.java
@@ -1,4 +1,4 @@
-package kr.hhplus.be.server.order;
+package kr.hhplus.be.server.order.domain;
 
 import jakarta.persistence.*;
 import kr.hhplus.be.server.product.Product;
