+++ src/main/java/kr/hhplus/be/server/order/dto/request/CreateOrderRequest.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.order.dto.request;
+
+public class CreateOrderRequest {
+}
