+++ src/main/java/kr/hhplus/be/server/order/dto/request/OrderItemRequest.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.order.dto.request;
+
+public class OrderItemRequest {
+}
