+++ src/main/java/kr/hhplus/be/server/order/dto/response/OrderResponse.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.order.dto.response;
+
+public class OrderResponse {
+}
