+++ src/main/java/kr/hhplus/be/server/order/infrastructure/OrderRepository.java
@@ -1,0 +1,7 @@
+package kr.hhplus.be.server.order.infrastructure;
+
+import kr.hhplus.be.server.order.domain.Order;
+import org.springframework.data.jpa.repository.JpaRepository;
+
+public interface OrderRepository extends JpaRepository<Order, Long> {
+}
