+++ src/main/java/kr/hhplus/be/server/order/presentation/OrderController.java
@@ -1,0 +1,4 @@
+package kr.hhplus.be.server.order.presentation;
+
+public class OrderController {
+}
